--- a/.pre-commit-config.yaml
+++ b/.pre-commit-config.yaml
@@ -26,6 +26,7 @@ repos:
               .*\.(json|prisma|svg)
               |.*pnpm-lock.yaml
               |.*deno.lock
+              |.*.yaml
           )$
         args:
           - "-L"