--- a/py/src/braintrust/wrappers/cassettes/test_anthropic_client_error.yaml
+++ b/py/src/braintrust/wrappers/cassettes/test_anthropic_client_error.yaml
@@ -1,0 +1,80 @@
+interactions:
+  - request:
+      body: '{"max_tokens":999,"messages":[{"role":"user","content":"who are you?"}],"model":"there-is-no-such-model"}'
+      headers:
+        accept:
+          - application/json
+        accept-encoding:
+          - gzip, deflate
+        anthropic-version:
+          - "2023-06-01"
+        connection:
+          - keep-alive
+        content-length:
+          - "105"
+        content-type:
+          - application/json
+        host:
+          - api.anthropic.com
+        user-agent:
+          - Anthropic/Python 0.52.1
+        x-stainless-arch:
+          - arm64
+        x-stainless-async:
+          - "false"
+        x-stainless-lang:
+          - python
+        x-stainless-os:
+          - MacOS
+        x-stainless-package-version:
+          - 0.52.1
+        x-stainless-read-timeout:
+          - "600"
+        x-stainless-retry-count:
+          - "0"
+        x-stainless-runtime:
+          - CPython
+        x-stainless-runtime-version:
+          - 3.13.3
+        x-stainless-timeout:
+          - "600"
+      method: POST
+      uri: https://api.anthropic.com/v1/messages
+    response:
+      body:
+        string: !!binary |
+          H4sIAAAAAAAAAwAAAP//qlYqqSxIVbJSSi0qyi9S0oHSVtUw8bz8kvi0/NK8lHiYitzU4uLEdJBc
+          bn5Kao6VQklGalGqbmaxbl6+bnFpcoYuWFypthYAAAD//wMAowdmMl0AAAA=
+      headers:
+        CF-RAY:
+          - 9472c8a12ade1b53-EWR
+        Connection:
+          - keep-alive
+        Content-Encoding:
+          - gzip
+        Content-Type:
+          - application/json
+        Date:
+          - Thu, 29 May 2025 03:07:49 GMT
+        Server:
+          - cloudflare
+        Transfer-Encoding:
+          - chunked
+        X-Robots-Tag:
+          - none
+        anthropic-organization-id:
+          - 27796668-7351-40ac-acc4-024aee8995a5
+        cf-cache-status:
+          - DYNAMIC
+        request-id:
+          - req_011CPb5rA13qiZ2bwi1DTVZp
+        strict-transport-security:
+          - max-age=31536000; includeSubDomains; preload
+        via:
+          - 1.1 google
+        x-should-retry:
+          - "false"
+      status:
+        code: 404
+        message: Not Found
+version: 1