This PR adds `pytest-vcr` to save http requests to openai and anthropic on disk and re-use them in tests, which speeds things up a lot and reduces flakiness. To balance the loss of testing reality, we always call the API once per CI run, but not for every python/os/library version. We can revisit this policy at some point, but i think it's ok for now.

```
pytest src/braintrust/wrappers/test_openai.py  0.87s user 0.14s system 97% cpu 1.032 total
pytest src/braintrust/wrappers/test_openai.py --disable-vcr  1.08s user 0.22s system 5% cpu 23.921 total
```

Installing libraries into `venvs` is the main bottleneck now and `uv` is surprisingly a lot faster. 
```
nox -s test_braintrust_core -db uv  1.18s user 0.58s system 43% cpu 4.042 total
nox -s test_braintrust_core -db venv  3.37s user 1.02s system 57% cpu 7.659 total
```

Before:
```
build (3.9, windows-latest) | 9m 52s
build (3.11, ubuntu-latest) | 6m 11s |  
```

After:

```
build (3.9, windows-latest) |  4m 52s |  
build (3.11, ubuntu-latest) | 2m 55s |  
```